requirements.txt
asgiref==3.5.0
botocore==1.29.57
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.2.1
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
gunicorn==20.1.0
jmespath==0.10.0
psycopg2-binary==2.9.5
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
